+++ tests/extendedfieldelement_tests.cc
@@ -56,14 +56,14 @@ TEST_F(ExtendedFieldElementFixtureGF8, GetField) {
   ASSERT_EQ(efe.getField(), ef);
 }
 
-TEST_F(ExtendedFieldElementFixtureGF8, Assignment) {
+TEST_F(ExtendedFieldElementFixtureGF8, Assign) {
   rslib::ExtendedFieldElement first(1, ef);
   rslib::ExtendedFieldElement second(2, ef);
   first = second;
   ASSERT_EQ(first.getValue(), 2u);
 }
 
-TEST_F(ExtendedFieldElementFixtureGF8, AssignmentToElementFromDifferentField) {
+TEST_F(ExtendedFieldElementFixtureGF8, AssignToElementFromDifferentField) {
   // If one extended field element is assigned to the second one then they
   // have to be in the same field.
   rslib::ExtendedFieldElement first(1, ef);
@@ -195,69 +195,44 @@ TEST_F(ExtendedFieldElementFixtureGF9, PolynomialRepresentationx) {
 
 TEST_F(ExtendedFieldElementFixtureGF8, ZechLogarithms) {
   ASSERT_EQ(rslib::ExtendedFieldElement(0, ef).getZechLogarithm(),
-            rslib::ExtendedField::ZechLogarithm(0));
+            rslib::ZechLogarithm(0));
   ASSERT_EQ(rslib::ExtendedFieldElement(1, ef).getZechLogarithm(),
-            rslib::ExtendedField::ZechLogarithm(
+            rslib::ZechLogarithm(
               std::numeric_limits<unsigned int>::max()));
   ASSERT_EQ(rslib::ExtendedFieldElement(2, ef).getZechLogarithm(),
-            rslib::ExtendedField::ZechLogarithm(3));
+            rslib::ZechLogarithm(3));
   ASSERT_EQ(rslib::ExtendedFieldElement(3, ef).getZechLogarithm(),
-            rslib::ExtendedField::ZechLogarithm(6));
+            rslib::ZechLogarithm(6));
   ASSERT_EQ(rslib::ExtendedFieldElement(4, ef).getZechLogarithm(),
-            rslib::ExtendedField::ZechLogarithm(1));
+            rslib::ZechLogarithm(1));
   ASSERT_EQ(rslib::ExtendedFieldElement(5, ef).getZechLogarithm(),
-            rslib::ExtendedField::ZechLogarithm(5));
+            rslib::ZechLogarithm(5));
   ASSERT_EQ(rslib::ExtendedFieldElement(6, ef).getZechLogarithm(),
-            rslib::ExtendedField::ZechLogarithm(4));
+            rslib::ZechLogarithm(4));
   ASSERT_EQ(rslib::ExtendedFieldElement(7, ef).getZechLogarithm(),
-            rslib::ExtendedField::ZechLogarithm(2));
+            rslib::ZechLogarithm(2));
 }
 
 TEST_F(ExtendedFieldElementFixtureGF9, ZechLogarithms) {
   ASSERT_EQ(rslib::ExtendedFieldElement(0, ef).getZechLogarithm(),
-            rslib::ExtendedField::ZechLogarithm(0));
+            rslib::ZechLogarithm(0));
   ASSERT_EQ(rslib::ExtendedFieldElement(1, ef).getZechLogarithm(),
-            rslib::ExtendedField::ZechLogarithm(8));
+            rslib::ZechLogarithm(8));
   ASSERT_EQ(rslib::ExtendedFieldElement(2, ef).getZechLogarithm(),
-            rslib::ExtendedField::ZechLogarithm(7));
+            rslib::ZechLogarithm(7));
   ASSERT_EQ(rslib::ExtendedFieldElement(3, ef).getZechLogarithm(),
-            rslib::ExtendedField::ZechLogarithm(3));
+            rslib::ZechLogarithm(3));
   ASSERT_EQ(rslib::ExtendedFieldElement(4, ef).getZechLogarithm(),
-            rslib::ExtendedField::ZechLogarithm(5));
+            rslib::ZechLogarithm(5));
   ASSERT_EQ(rslib::ExtendedFieldElement(5, ef).getZechLogarithm(),
-            rslib::ExtendedField::ZechLogarithm(
+            rslib::ZechLogarithm(
               std::numeric_limits<unsigned int>::max()));
   ASSERT_EQ(rslib::ExtendedFieldElement(6, ef).getZechLogarithm(),
-            rslib::ExtendedField::ZechLogarithm(2));
+            rslib::ZechLogarithm(2));
   ASSERT_EQ(rslib::ExtendedFieldElement(7, ef).getZechLogarithm(),
-            rslib::ExtendedField::ZechLogarithm(1));
+            rslib::ZechLogarithm(1));
   ASSERT_EQ(rslib::ExtendedFieldElement(8, ef).getZechLogarithm(),
-            rslib::ExtendedField::ZechLogarithm(6));
-}
-
-TEST_F(ExtendedFieldElementFixtureGF8, ZechLogarithmsPrint) {
-  rslib::ExtendedField::ZechLogarithm zechLogarithm
-      = rslib::ExtendedFieldElement(0, ef).getZechLogarithm();
-  std::ostringstream iss;
-  iss << zechLogarithm;
-  ASSERT_EQ(iss.str(), "0");
-}
-
-TEST_F(ExtendedFieldElementFixtureGF8, ZechLogarithmsPrintMinusInfinity) {
-  rslib::ExtendedField::ZechLogarithm zechLogarithm
-      = rslib::ExtendedFieldElement(1, ef).getZechLogarithm();
-  std::ostringstream iss;
-  iss << zechLogarithm;
-  ASSERT_EQ(iss.str(), "-INF");
-}
-
-TEST_F(ExtendedFieldElementFixtureGF8, ZechLogarithmsIsMinusInfinity) {
-  rslib::ExtendedField::ZechLogarithm zechLogarithmMinusInfinity
-      = rslib::ExtendedFieldElement(1, ef).getZechLogarithm();
-  ASSERT_TRUE(zechLogarithmMinusInfinity.isMinusInfinity());
-  rslib::ExtendedField::ZechLogarithm zechLogarithmNotMinusInfinity
-      = rslib::ExtendedFieldElement(0, ef).getZechLogarithm();
-  ASSERT_FALSE(zechLogarithmNotMinusInfinity.isMinusInfinity());
+            rslib::ZechLogarithm(6));
 }
 
 TEST_F(ExtendedFieldElementFixtureGF8, Print) {
@@ -277,3 +252,179 @@ TEST_F(ExtendedFieldElementFixtureGF8, Print) {
   iss4 << rslib::ExtendedFieldElement(3, ef);
   ASSERT_EQ(iss4.str(), "A^2");
 }
+
+TEST_F(ExtendedFieldElementFixtureGF8, Add) {
+  // A^3 = A + 1 in GF(8)
+  rslib::ExtendedFieldElement first(1, ef);
+  rslib::ExtendedFieldElement second(2, ef);
+  ASSERT_EQ(first + second, rslib::ExtendedFieldElement(4, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF8, AddWithZeroFirstArgument) {
+  rslib::ExtendedFieldElement first(3, ef);
+  rslib::ExtendedFieldElement second(0, ef);
+  ASSERT_EQ(first + second, rslib::ExtendedFieldElement(3, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF8, AddWithZeroSecondArgument) {
+  rslib::ExtendedFieldElement first(0, ef);
+  rslib::ExtendedFieldElement second(3, ef);
+  ASSERT_EQ(first + second, rslib::ExtendedFieldElement(3, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF8, AddWithAdditiveInverse) {
+  // A^2 + A^2 = 0 in GF(8)
+  rslib::ExtendedFieldElement first(3, ef);
+  rslib::ExtendedFieldElement second(3, ef);
+  ASSERT_EQ(first + second, rslib::ExtendedFieldElement(0, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF8, AdditiveInverse) {
+  // A^2 = -A^2 in GF(8)
+  rslib::ExtendedFieldElement element(3, ef);
+  ASSERT_EQ(-element, element);
+}
+
+TEST_F(ExtendedFieldElementFixtureGF8, AdditiveInverseOfZero) {
+  // 0 = -0
+  rslib::ExtendedFieldElement element(0, ef);
+  ASSERT_EQ(-element, element);
+}
+
+TEST_F(ExtendedFieldElementFixtureGF9, Add) {
+  // A^7 = A + 1 in GF(9)
+  rslib::ExtendedFieldElement first(1, ef);
+  rslib::ExtendedFieldElement second(2, ef);
+  ASSERT_EQ(first + second, rslib::ExtendedFieldElement(8, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF9, AddWithZeroFirstArgument) {
+  rslib::ExtendedFieldElement first(0, ef);
+  rslib::ExtendedFieldElement second(3, ef);
+  ASSERT_EQ(first + second, rslib::ExtendedFieldElement(3, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF9, AddWithZeroSecondArgument) {
+  rslib::ExtendedFieldElement first(3, ef);
+  rslib::ExtendedFieldElement second(0, ef);
+  ASSERT_EQ(first + second, rslib::ExtendedFieldElement(3, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF9, AddWithAdditiveInverse) {
+  // A^4 + 1 = 0 in GF(9)
+  rslib::ExtendedFieldElement first(5, ef);
+  rslib::ExtendedFieldElement second(1, ef);
+  ASSERT_EQ(first + second, rslib::ExtendedFieldElement(0, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF9, AdditiveInverse) {
+  // -A^4 = 1 in GF(9)
+  rslib::ExtendedFieldElement element(5, ef);
+  ASSERT_EQ(-element, rslib::ExtendedFieldElement(1, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF9, AdditiveInverseOfZero) {
+  // 0 = -0
+  rslib::ExtendedFieldElement element(0, ef);
+  ASSERT_EQ(-element, element);
+}
+
+TEST_F(ExtendedFieldElementFixtureGF8, Subtract) {
+  // 1 = A^3 - A in GF(8)
+  rslib::ExtendedFieldElement first(4, ef);
+  rslib::ExtendedFieldElement second(2, ef);
+  ASSERT_EQ(first - second, rslib::ExtendedFieldElement(1, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF9, Subtract) {
+  // 1 = A^7 - A in GF(9)
+  rslib::ExtendedFieldElement first(8, ef);
+  rslib::ExtendedFieldElement second(2, ef);
+  ASSERT_EQ(first - second, rslib::ExtendedFieldElement(1, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF8, Multiply) {
+  // A^4 = A^3 * A in GF(8)
+  rslib::ExtendedFieldElement first(4, ef);
+  rslib::ExtendedFieldElement second(2, ef);
+  ASSERT_EQ(first * second, rslib::ExtendedFieldElement(5, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF8, MultiplyWithZeroFirstArgument) {
+  rslib::ExtendedFieldElement first(0, ef);
+  rslib::ExtendedFieldElement second(4, ef);
+  ASSERT_EQ(first * second, rslib::ExtendedFieldElement(0, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF8, MultiplyWithZeroSecondArgument) {
+  rslib::ExtendedFieldElement first(4, ef);
+  rslib::ExtendedFieldElement second(0, ef);
+  ASSERT_EQ(first * second, rslib::ExtendedFieldElement(0, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF9, Multiply) {
+  // A^4 = A^3 * A in GF(9)
+  rslib::ExtendedFieldElement first(4, ef);
+  rslib::ExtendedFieldElement second(2, ef);
+  ASSERT_EQ(first * second, rslib::ExtendedFieldElement(5, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF9, MultiplyWithZeroFirstArgument) {
+  rslib::ExtendedFieldElement first(0, ef);
+  rslib::ExtendedFieldElement second(4, ef);
+  ASSERT_EQ(first * second, rslib::ExtendedFieldElement(0, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF9, MultiplyWithZeroSecondArgument) {
+  rslib::ExtendedFieldElement first(4, ef);
+  rslib::ExtendedFieldElement second(0, ef);
+  ASSERT_EQ(first * second, rslib::ExtendedFieldElement(0, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF8, MultiplicativeInverse) {
+  // 1/A^6 = A in GF(8)
+  rslib::ExtendedFieldElement element(7, ef);
+  ASSERT_EQ(element.multiplicativeInverse(),
+            rslib::ExtendedFieldElement(2, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF8, MultiplicativeInverseOfZero) {
+  // There is no multiplicative inverse of zero.
+  rslib::ExtendedFieldElement element(0, ef);
+  ASSERT_THROW(element.multiplicativeInverse(),
+               rslib::ExtendedFieldElementException);
+}
+
+TEST_F(ExtendedFieldElementFixtureGF8, MultiplicativeInverseOfOne) {
+  // 1/1 = 1
+  rslib::ExtendedFieldElement element(1, ef);
+  ASSERT_EQ(element.multiplicativeInverse(), element);
+}
+
+TEST_F(ExtendedFieldElementFixtureGF9, MultiplicativeInverse) {
+  // 1/A^6 = A^2 in GF(9)
+  rslib::ExtendedFieldElement element(7, ef);
+  ASSERT_EQ(element.multiplicativeInverse(),
+            rslib::ExtendedFieldElement(3, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF8, Divide) {
+  // A^4 / A^3 = A in GF(8)
+  rslib::ExtendedFieldElement first(5, ef);
+  rslib::ExtendedFieldElement second(4, ef);
+  ASSERT_EQ(first / second, rslib::ExtendedFieldElement(2, ef));
+}
+
+TEST_F(ExtendedFieldElementFixtureGF8, DivideByZero) {
+  rslib::ExtendedFieldElement first(5, ef);
+  rslib::ExtendedFieldElement second(0, ef);
+  ASSERT_THROW(first / second, rslib::ExtendedFieldElementException);
+}
+
+TEST_F(ExtendedFieldElementFixtureGF9, Divide) {
+  // A^4 / A^3 = A in GF(9)
+  rslib::ExtendedFieldElement first(5, ef);
+  rslib::ExtendedFieldElement second(4, ef);
+  ASSERT_EQ(first / second, rslib::ExtendedFieldElement(2, ef));
+}
+